refactor: create brotli function, remove compress, decompress functions
